+++ emapp/src/injector.rs
@@ -4,8 +4,8 @@ pub struct Injector {
     pub pixel_format: wgpu::TextureFormat,
     pub window_device_pixel_ratio: f32,
     pub viewport_device_pixel_ratio: f32,
-    pub window_size: [u16; 2],
-    pub viewport_size: [u16; 2],
+    pub window_size: [u32; 2],
+    pub viewport_size: [u32; 2],
 }
 
 impl Injector {
